Fix "White bars" glitch when maximising the window (Windows 10)
On Windows 10, there is a graphical glitch that appears when maximising the frameless window.
On a single monitor setup, two white bars will appear on the top and left corner for one frame before disappearing.  On a dual-monitor setup, the bars will remain forever until you manually repaint.  
This patch fixes the graphical glitch by forcing a repaint when resizing the non-client rect.

We also do not need to repaint before restoring the window; I could not see the twinkle with or without the repaint.